IIS Crash if Query name not set
Crash if query name not set, happened after 2.0 upgrade and System.Json.Text changes.

Error: 'activity.Parameters' threw an exception of type 'System.Text.Json.JsonException'

The name of the query in the database is a different type, system.json.text Presumably, when editing the name cannot translate, this creates a null exception in the ContentForEachTaskDisplayDriver.
Fix/Workaround - Add a try catch for the exception.  This allows the user to